feat(mobile): Phase-0 triple-target build infra (Slices 0–2) (#1021)
* feat(mobile): Phase-0 triple-target build infra (Slices 0–2)

Stand up the build-system scaffolding for the Android Phase-0 MVP without
touching any product C++. Three CI-validatable slices, all hardware-free:

- Slice 0/0b — triple-target the shared CORE_SOURCES. SmatchetMobile (SHARED,
  `if(ANDROID)`) joins SmatchetStandalone + SmatchetCore_DX12 off the one glob.
  A new compile-only gate, SmatchetCore_PosixCheck (STATIC EXCLUDE_FROM_ALL,
  guarded by SMATCHET_BUILD_POSIX_CORE_CHECK, default OFF), compiles the whole
  core under host Linux clang through the non-`_WIN32` `#else` branch of every
  platform `#ifdef`. A static archive only compiles — never links — so it
  exercises the portability surface with zero GLFW/GL/X11 link dependency.
  GLFW is fetched only `if(NOT ANDROID AND NOT IOS)`; ImGuiLib_Mobile is the
  Android-side imgui arm; the Standalone exe section is wrapped `if(NOT ANDROID)`.

- Slice 1 — TLS-backend spike. smatchet_prepare_cpr() forces the cpr OpenSSL
  backend + static libs under `if(ANDROID)` and widens CMAKE_FIND_ROOT_PATH_MODE_*
  to BOTH so find_package(OpenSSL) resolves a caller-supplied prebuilt OpenSSL
  outside the NDK sysroot (paths stay runtime-supplied via CI, never hardcoded).

- Slice 2 — empty `.so` link + two advisory CI jobs (not on the merge-gate):
  `mobile-posix-core-check` (host Linux clang, builds SmatchetCore_PosixCheck;
  installs xorg-dev/libgl1-mesa-dev so glfw *configures* but never *builds*) and
  `mobile-android-ndk` (NDK 26.3.11579264, API 24; source-builds OpenSSL 3.5.6
  per-ABI with a pinned SHA-256 + actions/cache, then configures + links
  SmatchetMobile against it).

New CMake presets posix-core-check + android-ndk-arm64 (configure + build) drive
both jobs. Every existing preset is byte-for-byte unaffected (new option defaults
OFF; new targets are EXCLUDE_FROM_ALL / `if(ANDROID)`).

Refs #1018. Plan: docs/plans/active/mobile-app-jql-mvp.md.


* fix(mobile): green the Slice 0-2 advisory CI + clear 2 CR findings

CI reds:
- TrackerHttpUtils.cpp: cast long->int32 for cpr ConnectTimeout/Timeout.
  LP64 `long` (Linux/Android) is 64-bit, so braced-init {long} into cpr's
  std::int32_t ctor narrows — a clang hard error, not a warning. Compiled on
  Windows only because LLP64 `long` is 32-bit. HTTP timeouts always fit int32.
- build-and-test.yml NDK install: `yes | sdkmanager --licenses` took EPIPE
  and tripped `set -o pipefail`; guard the pipeline with `|| true` and run the
  install pipe-free (the `test -d` check is the authoritative success gate).
- Both mobile build steps now pass `-- -k 0` (ninja keep-going) so a fresh
  portability sweep enumerates every failing TU in one run.

CR findings:
- DRY: drop CPR_FORCE_OPENSSL_BACKEND / OPENSSL_USE_STATIC_LIBS /
  CMAKE_FIND_ROOT_PATH_MODE_* from the android-ndk-arm64 preset. They are
  FORCE-set in smatchet_prepare_cpr() (runs before add_subdirectory(cpr)),
  which is now documented as the single authoritative home (covers raw
  `cmake` invocations too). Behavior-identical.
- Add a discoverable TODO(Slice-3) marker on SmatchetMobile's -Wl,-z,undefs.


* fix(mobile): portable std::string ctor in SubprocessCapture #else branch

The android NDK compile (Bionic libc++) hit:
  SubprocessCapture.cpp:475: no matching constructor for std::string
in the manual-environ-purge #else branch. environ[0] is char*, eq is
const char* (strchr result), and the two-iterator basic_string(first, last)
template needs matching pointer types — Bionic's libc++ rejects the mix.
Host glibc never compiled this branch (it takes the clearenv() #if side via
__GLIBC__), which is why the posix-core-check gate could not catch it. Switch
to the (ptr, length) ctor, unambiguous on every stdlib.

Found by the advisory android-ndk job once -- -k 0 let it compile past the
LP64 timeout fix; it was the only remaining failing TU (reached [405/407]).


---------

+++ Source/Core/src/Tracker/TrackerHttpUtils.cpp
@@ -128,8 +128,13 @@ cpr::Response TrackerGetLogged(const char* clientName, const std::string& url, c
 cpr::Response TrackerGetLogged(const char* clientName, const std::string& url, const cpr::Header& headers,
                                long connectTimeoutMs, long overallTimeoutMs) {
     cpr::Redirect redirect(true, true);
-    cpr::Response response = cpr::Get(cpr::Url{url}, headers, redirect, cpr::ConnectTimeout{connectTimeoutMs},
-                                      cpr::Timeout{overallTimeoutMs});
+    // cpr's ConnectTimeout/Timeout take std::int32_t. These params are `long`, which is
+    // 64-bit on LP64 (Linux/Android), so a braced-init {long} narrows — ill-formed under
+    // clang (hard error, not a warning). It compiles on Windows only because LLP64 `long`
+    // is 32-bit. Cast explicitly; HTTP timeouts in ms always fit int32 (<= ~24.8 days).
+    cpr::Response response = cpr::Get(cpr::Url{url}, headers, redirect,
+                                      cpr::ConnectTimeout{static_cast<std::int32_t>(connectTimeoutMs)},
+                                      cpr::Timeout{static_cast<std::int32_t>(overallTimeoutMs)});
     NetworkUsageTracker::Instance().Record(HttpTrafficKind::Tracker, NetworkUsageTracker::kEstimatedGetUploadBytes,
                                            response);
     LogTrackerHttpResult(clientName, "GET", url, response);
